% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_theme.R
\name{wb_col}
\alias{wb_col}
\title{Function to extract WB colors as hex codes}
\usage{
wb_col(...)
}
\arguments{
\item{...}{Character names of drsimonj_colors}
}
\description{
Function to extract WB colors as hex codes
}
